+++ packages/sandbox-local/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @effect-agent/sandbox-local
 
+## 0.1.0-beta.62
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @effect-agent/sandbox@0.1.0-beta.62
+
+## 0.1.0-beta.61
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @effect-agent/sandbox@0.1.0-beta.61
+
+## 0.1.0-beta.60
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @effect-agent/sandbox@0.1.0-beta.60
+
 ## 0.1.0-beta.59
 
 ### Patch Changes